fix: 🐛 split correctly numbered lists
Issues: #35
