+++ fixtures/json/multiple-paragraphs.json
@@ -108,7 +108,7 @@
           "children": [
             {
               "type": "paragraph",
-              "len": 104,
+              "len": 103,
               "children": [
                 {
                   "type": "text",
@@ -116,116 +116,98 @@
                   "value": "Supports bare module specifiers: This is the npm standard, and modern tooling like bundlers support it."
                 }
               ]
-            },
+            }
+          ]
+        },
+        {
+          "type": "listItem",
+          "loose": false,
+          "checked": null,
+          "children": [
             {
-              "type": "list",
-              "len": 325,
+              "type": "paragraph",
+              "len": 150,
               "children": [
                 {
-                  "type": "listItem",
-                  "loose": false,
-                  "checked": null,
+                  "type": "text",
+                  "len": 17,
+                  "value": "Supports dynamic "
+                },
+                {
+                  "type": "inlineCode",
+                  "len": 10,
+                  "value": "import()",
+                  "wrap": "`"
+                },
+                {
+                  "type": "text",
+                  "len": 12,
+                  "value": ": This is a "
+                },
+                {
+                  "type": "link",
+                  "len": 67,
                   "children": [
                     {
-                      "type": "paragraph",
-                      "len": 151,
-                      "children": [
-                        {
-                          "type": "text",
-                          "len": 17,
-                          "value": "Supports dynamic "
-                        },
-                        {
-                          "type": "inlineCode",
-                          "len": 10,
-                          "value": "import()",
-                          "wrap": "`"
-                        },
-                        {
-                          "type": "text",
-                          "len": 12,
-                          "value": ": This is a "
-                        },
-                        {
-                          "type": "link",
-                          "len": 67,
-                          "children": [
-                            {
-                              "type": "text",
-                              "len": 16,
-                              "value": "stage 3 proposal"
-                            }
-                          ],
-                          "url": "https://github.com/tc39/proposal-dynamic-import"
-                        },
-                        {
-                          "type": "text",
-                          "len": 44,
-                          "value": " which allows dynamically importing modules."
-                        }
-                      ]
-                    },
-                    {
-                      "type": "list",
-                      "len": 170,
-                      "children": [
-                        {
-                          "type": "listItem",
-                          "loose": false,
-                          "checked": null,
-                          "children": [
-                            {
-                              "type": "paragraph",
-                              "len": 166,
-                              "children": [
-                                {
-                                  "type": "text",
-                                  "len": 9,
-                                  "value": "Supports "
-                                },
-                                {
-                                  "type": "inlineCode",
-                                  "len": 17,
-                                  "value": "import.meta.url",
-                                  "wrap": "`"
-                                },
-                                {
-                                  "type": "text",
-                                  "len": 12,
-                                  "value": ": This is a "
-                                },
-                                {
-                                  "type": "link",
-                                  "len": 64,
-                                  "children": [
-                                    {
-                                      "type": "text",
-                                      "len": 16,
-                                      "value": "stage 3 proposal"
-                                    }
-                                  ],
-                                  "url": "https://github.com/tc39/proposal-import-meta"
-                                },
-                                {
-                                  "type": "text",
-                                  "len": 63,
-                                  "value": " which allows modules to read their location (filepath or URL)."
-                                }
-                              ]
-                            }
-                          ]
-                        }
-                      ],
-                      "ordered": true,
-                      "start": 1,
-                      "loose": false
+                      "type": "text",
+                      "len": 16,
+                      "value": "stage 3 proposal"
                     }
-                  ]
+                  ],
+                  "url": "https://github.com/tc39/proposal-dynamic-import"
+                },
+                {
+                  "type": "text",
+                  "len": 44,
+                  "value": " which allows dynamically importing modules."
                 }
-              ],
-              "ordered": true,
-              "start": 1,
-              "loose": false
+              ]
+            }
+          ]
+        },
+        {
+          "type": "listItem",
+          "loose": false,
+          "checked": null,
+          "children": [
+            {
+              "type": "paragraph",
+              "len": 166,
+              "children": [
+                {
+                  "type": "text",
+                  "len": 9,
+                  "value": "Supports "
+                },
+                {
+                  "type": "inlineCode",
+                  "len": 17,
+                  "value": "import.meta.url",
+                  "wrap": "`"
+                },
+                {
+                  "type": "text",
+                  "len": 12,
+                  "value": ": This is a "
+                },
+                {
+                  "type": "link",
+                  "len": 64,
+                  "children": [
+                    {
+                      "type": "text",
+                      "len": 16,
+                      "value": "stage 3 proposal"
+                    }
+                  ],
+                  "url": "https://github.com/tc39/proposal-import-meta"
+                },
+                {
+                  "type": "text",
+                  "len": 63,
+                  "value": " which allows modules to read their location (filepath or URL)."
+                }
+              ]
             }
           ]
         }
