+++ fixtures/md/list-ordered.md
@@ -1,0 +1,2 @@
+1. Item 1
+2. Item 2
