+++ src/__tests__/block.spec.ts
@@ -531,346 +531,6 @@ trololo`);
         });
     });
 
-    describe('list', () => {
-        it('works', () => {
-            const parser = create();
-            const ast = parser.tokenizeBlock(`- foo`);
-
-            expect(ast).toMatchObject({
-                type: 'root',
-                children: [
-                    {
-                        type: 'list',
-                        ordered: false,
-                        loose: false,
-                        len: 5,
-                        children: [
-                            {
-                                type: 'listItem',
-                                checked: null,
-                                loose: false,
-                                children: [
-                                    {
-                                        type: 'paragraph',
-                                        len: 3,
-                                        children: [
-                                            {
-                                                type: 'text',
-                                                len: 3,
-                                                value: 'foo',
-                                            },
-                                        ],
-                                    },
-                                ],
-                            },
-                        ],
-                        start: null,
-                    },
-                ],
-                len: 5,
-            });
-        });
-
-        it('supports all bullet markers', () => {
-            const result = {
-                type: 'root',
-                children: [
-                    {
-                        type: 'list',
-                        ordered: false,
-                        loose: false,
-                        len: 5,
-                        children: [
-                            {
-                                type: 'listItem',
-                                checked: null,
-                                loose: false,
-                                children: [
-                                    {
-                                        type: 'paragraph',
-                                        len: 3,
-                                        children: [
-                                            {
-                                                type: 'text',
-                                                len: 3,
-                                                value: 'foo',
-                                            },
-                                        ],
-                                    },
-                                ],
-                            },
-                        ],
-                        start: null,
-                    },
-                ],
-                len: 5,
-            };
-
-            const parser = create();
-
-            expect(parser.tokenizeBlock(`- foo`)).toMatchObject(result);
-            expect(parser.tokenizeBlock(`* foo`)).toMatchObject(result);
-            expect(parser.tokenizeBlock(`+ foo`)).toMatchObject(result);
-        });
-
-        it('supports ordered lists', () => {
-            const parser = create();
-            const ast = parser.tokenizeBlock(`1. foo`);
-
-            expect(ast).toMatchObject({
-                type: 'root',
-                children: [
-                    {
-                        type: 'list',
-                        ordered: true,
-                        loose: false,
-                        len: 6,
-                        children: [
-                            {
-                                type: 'listItem',
-                                checked: null,
-                                loose: false,
-                                children: [
-                                    {
-                                        type: 'paragraph',
-                                        len: 3,
-                                        children: [
-                                            {
-                                                type: 'text',
-                                                len: 3,
-                                                value: 'foo',
-                                            },
-                                        ],
-                                    },
-                                ],
-                            },
-                        ],
-                        start: 1,
-                    },
-                ],
-                len: 6,
-            });
-        });
-
-        it('reports loose items', () => {
-            const parser = create();
-            const ast = parser.tokenizeBlock(`- foo\n\n  bar`);
-
-            expect(ast).toMatchObject({
-                type: 'root',
-                children: [
-                    {
-                        type: 'list',
-                        len: 12,
-                        children: [
-                            {
-                                type: 'listItem',
-                                loose: true,
-                                checked: null,
-                                children: [
-                                    {
-                                        type: 'paragraph',
-                                        len: 5,
-                                        children: [
-                                            {
-                                                type: 'text',
-                                                len: 3,
-                                                value: 'foo',
-                                            },
-                                        ],
-                                    },
-                                    {
-                                        type: 'paragraph',
-                                        len: 3,
-                                        children: [
-                                            {
-                                                type: 'text',
-                                                len: 3,
-                                                value: 'bar',
-                                            },
-                                        ],
-                                    },
-                                ],
-                            },
-                        ],
-                        ordered: false,
-                        start: null,
-                        loose: true,
-                    },
-                ],
-                len: 12,
-            });
-        });
-
-        it('supports multiple items', () => {
-            const parser = create();
-            const ast = parser.tokenizeBlock(`- foo\n- bar\n- baz`);
-
-            expect(ast).toMatchObject({
-                type: 'root',
-                children: [
-                    {
-                        type: 'list',
-                        len: 17,
-                        children: [
-                            {
-                                type: 'listItem',
-                                loose: false,
-                                checked: null,
-                                children: [
-                                    {
-                                        type: 'paragraph',
-                                        len: 3,
-                                        children: [
-                                            {
-                                                type: 'text',
-                                                len: 3,
-                                                value: 'foo',
-                                            },
-                                        ],
-                                    },
-                                ],
-                            },
-                            {
-                                type: 'listItem',
-                                loose: false,
-                                checked: null,
-                                children: [
-                                    {
-                                        type: 'paragraph',
-                                        len: 3,
-                                        children: [
-                                            {
-                                                type: 'text',
-                                                len: 3,
-                                                value: 'bar',
-                                            },
-                                        ],
-                                    },
-                                ],
-                            },
-                            {
-                                type: 'listItem',
-                                loose: false,
-                                checked: null,
-                                children: [
-                                    {
-                                        type: 'paragraph',
-                                        len: 3,
-                                        children: [
-                                            {
-                                                type: 'text',
-                                                len: 3,
-                                                value: 'baz',
-                                            },
-                                        ],
-                                    },
-                                ],
-                            },
-                        ],
-                        ordered: false,
-                        start: null,
-                        loose: false,
-                    },
-                ],
-                len: 17,
-            });
-        });
-
-        it('supporst nested lists', () => {
-            const parser = create();
-            const ast = parser.tokenizeBlock(`- foo\n   - bar\n      - baz`);
-
-            expect(ast).toMatchObject({
-                type: 'root',
-                children: [
-                    {
-                        type: 'list',
-                        len: 26,
-                        children: [
-                            {
-                                type: 'listItem',
-                                loose: false,
-                                checked: null,
-                                children: [
-                                    {
-                                        type: 'paragraph',
-                                        len: 4,
-                                        children: [
-                                            {
-                                                type: 'text',
-                                                len: 3,
-                                                value: 'foo',
-                                            },
-                                        ],
-                                    },
-                                    {
-                                        type: 'list',
-                                        len: 13,
-                                        children: [
-                                            {
-                                                type: 'listItem',
-                                                loose: false,
-                                                checked: null,
-                                                children: [
-                                                    {
-                                                        type: 'paragraph',
-                                                        len: 4,
-                                                        children: [
-                                                            {
-                                                                type: 'text',
-                                                                len: 3,
-                                                                value: 'bar',
-                                                            },
-                                                        ],
-                                                    },
-                                                    {
-                                                        type: 'list',
-                                                        len: 5,
-                                                        children: [
-                                                            {
-                                                                type: 'listItem',
-                                                                loose: false,
-                                                                checked: null,
-                                                                children: [
-                                                                    {
-                                                                        type: 'paragraph',
-                                                                        len: 3,
-                                                                        children: [
-                                                                            {
-                                                                                type: 'text',
-                                                                                len: 3,
-                                                                                value: 'baz',
-                                                                            },
-                                                                        ],
-                                                                    },
-                                                                ],
-                                                            },
-                                                        ],
-                                                        ordered: false,
-                                                        start: null,
-                                                        loose: false,
-                                                    },
-                                                ],
-                                            },
-                                        ],
-                                        ordered: false,
-                                        start: null,
-                                        loose: false,
-                                    },
-                                ],
-                            },
-                        ],
-                        ordered: false,
-                        start: null,
-                        loose: false,
-                    },
-                ],
-                len: 26,
-            });
-        });
-    });
-
     describe('table', () => {
         it('works', () => {
             const table = `| Table |
